Add tests for task 2

+++ tests/test_console.py
@@ -27,17 +27,17 @@ class TestConsole(unittest.TestCase):
         """setup for the test"""
         cls.consol = HBNBCommand()
 
-    @classmethod
-    def teardown(cls):
-        """at the end of the test this will tear it down"""
-        del cls.consol
-
-    def tearDown(self):
-        """Remove temporary file (file.json) created as a result"""
-        try:
+    def setUp(self):
+        """Sets up test cases."""
+        if os.path.isfile("file.json"):
             os.remove("file.json")
-        except Exception:
-            pass
+        self.resetStorage()
+
+    def resetStorage(self):
+        """Resets FileStorage data."""
+        FileStorage._FileStorage__objects = {}
+        if os.path.isfile(FileStorage._FileStorage__file_path):
+            os.remove(FileStorage._FileStorage__file_path)
 
     def test_pep8_console(self):
         """Pep8 console.py"""
@@ -88,6 +88,52 @@ class TestConsole(unittest.TestCase):
             self.consol.onecmd("all User")
             self.assertEqual(
                 "[[User]", f.getvalue()[:7])
+
+    def test_create_v2(self):
+        """Test create command with parameters."""
+        with patch('sys.stdout', new=StringIO()) as f:
+            self.consol.onecmd('create State name="California"')
+            id = f.getvalue()[:-1]
+            self.assertEqual(len(id), 36)
+        with patch('sys.stdout', new=StringIO()) as f:
+            self.consol.onecmd("show State {}".format(id))
+            self.assertTrue("'name': 'California'" in f.getvalue())
+            self.assertEqual(
+                "[State]", f.getvalue()[:7])
+        with patch('sys.stdout', new=StringIO()) as f:
+            self.consol.onecmd('create City name="San_Francisco"')
+            id = f.getvalue()[:-1]
+            self.assertEqual(len(id), 36)
+        with patch('sys.stdout', new=StringIO()) as f:
+            self.consol.onecmd("show City {}".format(id))
+            self.assertTrue("'name': 'San Francisco'" in f.getvalue())
+        with patch('sys.stdout', new=StringIO()) as f:
+            self.consol.onecmd('create Place latitude=7.89')
+            id = f.getvalue()[:-1]
+            self.assertEqual(len(id), 36)
+        with patch('sys.stdout', new=StringIO()) as f:
+            self.consol.onecmd("show Place {}".format(id))
+            self.assertTrue("'latitude': 7.89" in f.getvalue())
+        with patch('sys.stdout', new=StringIO()) as f:
+            self.consol.onecmd('create Place max_guest=5')
+            id = f.getvalue()[:-1]
+            self.assertEqual(len(id), 36)
+        with patch('sys.stdout', new=StringIO()) as f:
+            self.consol.onecmd("show Place {}".format(id))
+            self.assertTrue("'max_guest': 5" in f.getvalue())
+
+    def test_create_v2_params(self):
+        """Test create command with several parameters"""
+        with patch('sys.stdout', new=StringIO()) as f:
+            self.consol.onecmd('create User email="ilovetim@google.com" password="timisboss" first_name="Farrukh" last_name')
+            id = f.getvalue()[:-1]
+            self.assertEqual(len(id), 36)
+        with patch('sys.stdout', new=StringIO()) as f:
+            self.consol.onecmd("show User {}".format(id))
+            self.assertTrue("'email': 'ilovetim@google.com'" in f.getvalue())
+            self.assertTrue("'password': 'timisboss'" in f.getvalue())
+            self.assertTrue("'first_name': 'Farrukh'" in f.getvalue())
+
 
     def test_show(self):
         """Test show command inpout"""
@@ -155,15 +201,14 @@ class TestConsole(unittest.TestCase):
             self.assertEqual(
                 "** no instance found **\n", f.getvalue())
         with patch('sys.stdout', new=StringIO()) as f:
-            self.consol.onecmd("all User")
-            obj = f.getvalue()
-        my_id = obj[obj.find('(')+1:obj.find(')')]
+            self.consol.onecmd("create User")
+            id = f.getvalue()[:-1]
         with patch('sys.stdout', new=StringIO()) as f:
-            self.consol.onecmd("update User " + my_id)
+            self.consol.onecmd("update User " + id)
             self.assertEqual(
                 "** attribute name missing **\n", f.getvalue())
         with patch('sys.stdout', new=StringIO()) as f:
-            self.consol.onecmd("update User " + my_id + " Name")
+            self.consol.onecmd("update User " + id + " Name")
             self.assertEqual(
                 "** value missing **\n", f.getvalue())
 
@@ -209,7 +254,7 @@ class TestConsole(unittest.TestCase):
             self.assertEqual(
                 "** no instance found **\n", f.getvalue())
 
-    def test_update(self):
+    def test_update_error(self):
         """Test alternate destroy command inpout"""
         with patch('sys.stdout', new=StringIO()) as f:
             self.consol.onecmd("sldkfjsl.update()")
@@ -220,15 +265,14 @@ class TestConsole(unittest.TestCase):
             self.assertEqual(
                 "** no instance found **\n", f.getvalue())
         with patch('sys.stdout', new=StringIO()) as f:
-            self.consol.onecmd("all User")
-            obj = f.getvalue()
-        my_id = obj[obj.find('(')+1:obj.find(')')]
+            self.consol.onecmd("create User")
+            id = f.getvalue()[:-1]
         with patch('sys.stdout', new=StringIO()) as f:
-            self.consol.onecmd("User.update(" + my_id + ")")
+            self.consol.onecmd("User.update(" + id + ")")
             self.assertEqual(
                 "** attribute name missing **\n", f.getvalue())
         with patch('sys.stdout', new=StringIO()) as f:
-            self.consol.onecmd("User.update(" + my_id + ", name)")
+            self.consol.onecmd("User.update(" + id + ", name)")
             self.assertEqual(
                 "** value missing **\n", f.getvalue())
